+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Arbimon Release Notes
 
-## v3.0.33 - June 01, 2021
+## v3.0.33 - July XX, 2021
 
 Performance improvements:
 
 - CE-988 The datime_local column is removed
+- CE-1143 Update "Unvalidated", "All", "Present", "Not present", "200 top scores per site", "Score per site" filters on Pattern Matching details page to return results per site
+- CE-1158 update PM filters logic so "Best per Site" and "Best per Site, Day" return results by 20 sites batches
 
 ## v3.0.32 - June 26, 2021